+++ packages/chain-adapters/src/bitcoin/BitcoinChainAdapter.test.ts
@@ -8,7 +8,7 @@
 
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
 import { NativeAdapterArgs, NativeHDWallet } from '@shapeshiftoss/hdwallet-native'
-import { BIP32Params, chainAdapters, ChainTypes, UtxoAccountType } from '@shapeshiftoss/types'
+import { BIP44Params, chainAdapters, ChainTypes, UtxoAccountType } from '@shapeshiftoss/types'
 import * as unchained from '@shapeshiftoss/unchained-client'
 
 import * as bitcoin from './BitcoinChainAdapter'
@@ -159,6 +159,7 @@ describe('BitcoinChainAdapter', () => {
   describe('getAccount', () => {
     it('should return account info for a specified address', async () => {
       args.providers.http = {
+        getInfo: jest.fn().mockResolvedValue({ data: { network: 'mainnet' } }),
         getAccount: jest.fn().mockResolvedValue({
           data: {
             pubkey: '1EjpFGTWJ9CGRJUMA3SdQSdigxM31aXAFx',
@@ -179,6 +180,7 @@ describe('BitcoinChainAdapter', () => {
 
     it('should throw for an unspecified address', async () => {
       args.providers.http = {
+        getInfo: jest.fn().mockResolvedValue({ data: { network: 'mainnet' } }),
         getAccount: jest.fn<any, any>().mockResolvedValue({
           pubkey: '1EjpFGTWJ9CGRJUMA3SdQSdigxM31aXAFx',
           balance: '0'
@@ -261,6 +263,7 @@ describe('BitcoinChainAdapter', () => {
       const wallet: any = await getWallet()
 
       args.providers.http = {
+        getInfo: jest.fn().mockResolvedValue({ data: { network: 'mainnet' } }),
         getUtxos: jest.fn<any, any>().mockResolvedValue(getUtxosMockResponse),
         getTransaction: jest.fn<any, any>().mockResolvedValue(getTransactionMockResponse),
         getAccount: jest.fn().mockResolvedValue(getAccountMockResponse),
@@ -269,7 +272,7 @@ describe('BitcoinChainAdapter', () => {
 
       const adapter = new bitcoin.ChainAdapter(args)
 
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         purpose: 84,
         coinType: 0,
         accountNumber: 0,
@@ -277,7 +280,7 @@ describe('BitcoinChainAdapter', () => {
       }
 
       const txInput: chainAdapters.BuildSendTxInput<ChainTypes.Bitcoin> = {
-        bip32Params,
+        bip44Params,
         to: 'bc1qppzsgs9pt63cx9x994wf4e3qrpta0nm6htk9v4',
         value: '400',
         wallet,
@@ -329,6 +332,7 @@ describe('BitcoinChainAdapter', () => {
       const wallet: any = await getWallet()
 
       args.providers.http = {
+        getInfo: jest.fn().mockResolvedValue({ data: { network: 'mainnet' } }),
         getUtxos: jest.fn<any, any>().mockResolvedValue(getUtxosMockResponse),
         getTransaction: jest.fn<any, any>().mockResolvedValue(getTransactionMockResponse),
         getAccount: jest.fn().mockResolvedValue(getAccountMockResponse),
@@ -337,7 +341,7 @@ describe('BitcoinChainAdapter', () => {
 
       const adapter = new bitcoin.ChainAdapter(args)
 
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         purpose: 84,
         coinType: 0,
         accountNumber: 0,
@@ -345,7 +349,7 @@ describe('BitcoinChainAdapter', () => {
       }
 
       const txInput: chainAdapters.BuildSendTxInput<ChainTypes.Bitcoin> = {
-        bip32Params,
+        bip44Params,
         to: 'bc1qppzsgs9pt63cx9x994wf4e3qrpta0nm6htk9v4',
         value: '400',
         wallet,
@@ -411,7 +415,7 @@ describe('BitcoinChainAdapter', () => {
     it("should return a p2pkh address for valid derivation root path parameters (m/44'/0'/0'/0/0)", async () => {
       const wallet: HDWallet = await getWallet()
       const adapter = new bitcoin.ChainAdapter(args)
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         coinType: 0,
         purpose: 44,
         accountNumber: 0,
@@ -420,7 +424,7 @@ describe('BitcoinChainAdapter', () => {
       }
 
       const addr: string | undefined = await adapter.getAddress({
-        bip32Params,
+        bip44Params,
         wallet,
         accountType: UtxoAccountType.P2pkh
       })
@@ -430,7 +434,7 @@ describe('BitcoinChainAdapter', () => {
     it("should return a valid p2pkh address for the first receive index path (m/44'/0'/0'/0/1)", async () => {
       const wallet: HDWallet = await getWallet()
       const adapter = new bitcoin.ChainAdapter(args)
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         coinType: 0,
         purpose: 44,
         accountNumber: 0,
@@ -438,7 +442,7 @@ describe('BitcoinChainAdapter', () => {
         isChange: false
       }
       const addr: string | undefined = await adapter.getAddress({
-        bip32Params,
+        bip44Params,
         wallet,
         accountType: UtxoAccountType.P2pkh
       })
@@ -448,7 +452,7 @@ describe('BitcoinChainAdapter', () => {
     it("should return a valid p2pkh change address for the first receive index path (m/44'/0'/0'/1/0)", async () => {
       const wallet: HDWallet = await getWallet()
       const adapter = new bitcoin.ChainAdapter(args)
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         coinType: 0,
         purpose: 44,
         accountNumber: 0,
@@ -456,7 +460,7 @@ describe('BitcoinChainAdapter', () => {
         isChange: true
       }
       const addr: string | undefined = await adapter.getAddress({
-        bip32Params,
+        bip44Params,
         wallet,
         accountType: UtxoAccountType.P2pkh
       })
@@ -466,7 +470,7 @@ describe('BitcoinChainAdapter', () => {
     it("should return a valid p2pkh address at the 2nd account root path (m/44'/0'/1'/0/0)", async () => {
       const wallet: HDWallet = await getWallet()
       const adapter = new bitcoin.ChainAdapter(args)
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         coinType: 0,
         purpose: 44,
         accountNumber: 1,
@@ -474,7 +478,7 @@ describe('BitcoinChainAdapter', () => {
         isChange: false
       }
       const addr: string | undefined = await adapter.getAddress({
-        bip32Params,
+        bip44Params,
         wallet,
         accountType: UtxoAccountType.P2pkh
       })
@@ -484,7 +488,7 @@ describe('BitcoinChainAdapter', () => {
     it("should return a p2wpkh address for valid derivation root path parameters (m/84'/0'/0'/0/0)", async () => {
       const wallet: HDWallet = await getWallet()
       const adapter = new bitcoin.ChainAdapter(args)
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         coinType: 0,
         purpose: 84,
         accountNumber: 0,
@@ -492,7 +496,7 @@ describe('BitcoinChainAdapter', () => {
         index: 0
       }
       const addr: string | undefined = await adapter.getAddress({
-        bip32Params,
+        bip44Params,
         wallet,
         accountType: UtxoAccountType.SegwitNative
       })
@@ -502,7 +506,7 @@ describe('BitcoinChainAdapter', () => {
     it("should return a valid p2wpkh address for the first receive index path (m/84'/0'/0'/0/1)", async () => {
       const wallet: HDWallet = await getWallet()
       const adapter = new bitcoin.ChainAdapter(args)
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         coinType: 0,
         purpose: 84,
         accountNumber: 0,
@@ -510,7 +514,7 @@ describe('BitcoinChainAdapter', () => {
         isChange: false
       }
       const addr: string | undefined = await adapter.getAddress({
-        bip32Params,
+        bip44Params,
         wallet,
         accountType: UtxoAccountType.SegwitNative
       })
@@ -520,7 +524,7 @@ describe('BitcoinChainAdapter', () => {
     it("should return a valid p2wpkh change address for the first receive index path (m/44'/0'/0'/1/0)", async () => {
       const wallet: HDWallet = await getWallet()
       const adapter = new bitcoin.ChainAdapter(args)
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         coinType: 0,
         purpose: 84,
         accountNumber: 0,
@@ -528,7 +532,7 @@ describe('BitcoinChainAdapter', () => {
         isChange: true
       }
       const addr: string | undefined = await adapter.getAddress({
-        bip32Params,
+        bip44Params,
         wallet,
         accountType: UtxoAccountType.SegwitNative
       })
@@ -538,7 +542,7 @@ describe('BitcoinChainAdapter', () => {
     it("should return a valid p2wpkh address at the 2nd account root path (m/84'/0'/1'/0/0)", async () => {
       const wallet: HDWallet = await getWallet()
       const adapter = new bitcoin.ChainAdapter(args)
-      const bip32Params: BIP32Params = {
+      const bip44Params: BIP44Params = {
         coinType: 0,
         purpose: 84,
         accountNumber: 1,
@@ -546,7 +550,7 @@ describe('BitcoinChainAdapter', () => {
         isChange: false
       }
       const addr: string | undefined = await adapter.getAddress({
-        bip32Params,
+        bip44Params,
         wallet,
         accountType: UtxoAccountType.SegwitNative
       })